code/requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
scipy>=1.11
pyradiomics>=3.0
SimpleITK>=2.3
torch>=2.0
torchvision>=0.15
scikit-image>=0.21
Pillow>=10.0
matplotlib>=3.7
seaborn>=0.12
tqdm>=4.66
PyYAML>=6.0
jupyter>=1.0
